Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/Close-reason-unpaired-surrogates.any.worker-expected.txt


PASS Create WebSocket - Close the Connection - close(reason with unpaired surrogates) - connection should get opened 
FAIL Create WebSocket - Close the Connection - close(reason with unpaired surrogates) - connection should get closed assert_equals: reason replaced with replacement character expected "\ufffd" but got ""
